lib: Allow creating non-leaf (or hierarchical) HSM instances
A cluster of harts on a platform can have different platform ops
for managing due to difference in implementation. For example,
a platform having a big-cluster and litter-cluster.

These platforms can create a separte leaf HSM instance for each
hart cluster and have a non-leaf or hierarchical HSM instance to
manage all leaf-level or cluster-level HSM instances.

Signed-off-by: Anup Patel <apatel@ventanamicro.com>