lib: Introduce platform specific shared memory operations
To make life easy for users who will integrate librpmi with an
emulator or simulator (such as QEMU), we introduce platform specific
shared memory operations. The simple shared memory is still supported
but in the form of shared memory platform operations.

In the process, we also move shared memory related code to it's
own source file.

Signed-off-by: Anup Patel <apatel@ventanamicro.com>