lib: Add CPPC performance feedback fastchannel support
Each hart has performance feedback fastchannel entry
which is used to get the current frequency of a hart
as a feedback whenever there is a change in performacne
level for that hart. This feedback of frequency is
used by the power management software in OS to adjust the
next performance level request.

Signed-off-by: Rahul Pathak <rpathak@ventanamicro.com>