ci: add a simple github action that is run on pull-requests to main
Add a simple github action that runs on every pull-request to the `main`
branch.

This checks that librpmi and sources can be built and tests on Ubuntu and
macos.

The `fail-fast` test strategy is disabled, so that both build targets
can be run. If either of the build targets fail, then the job fails.

Additionally, since librpmi CI is quite fast, add a 2 minute timeout, since
(currently) an invalid service group will cause the implementation to spin
infinitely.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>